Починил удаление работников и файлов

+++ My_WebApp/Controllers/FilesController.cs
@@ -10,7 +10,7 @@ namespace My_WebApp.Controllers
 {
     [Authorize]
     [ApiController]
-    [Route("api/[controller]")]
+    [Route("[controller]")]
 
     public class FilesController : ControllerBase
     {
@@ -22,7 +22,7 @@ namespace My_WebApp.Controllers
             _context = new ApplicationContext();
         }
 
-        [HttpGet("download")]
+        [HttpPost("download")]
         [Authorize(Roles = "admin,manager")]
         public IActionResult DownloadFile([FromBody] DownloadFileRequestDto downloadFileRequestDto)
         {
@@ -49,8 +49,8 @@ namespace My_WebApp.Controllers
         {
             try
             {
-                var employee = _context.Employees.Include(e=>e.UserFiles)
-                    .FirstOrDefault(e=>e.Id == fileDto.EmployeeId);
+                var employee = _context.Employees.Include(e => e.UserFiles)
+                    .FirstOrDefault(e => e.Id == fileDto.EmployeeId);
                 if (employee == null)
                 {
                     return BadRequest("User not found");
@@ -75,28 +75,43 @@ namespace My_WebApp.Controllers
                     SystemName = systemFileName
                 });
                 _context.SaveChanges();
+
                 return Ok();
             }
             catch (Exception ex)
             {
                 return BadRequest(ex.Message);
             }
+
         }
 
 
         [HttpDelete("delete")]
         [Authorize(Roles = "admin,manager")]
-        public IActionResult DeleteFile(string systemName)
+        public IActionResult DeleteFile(int id)
         {
             try
             {
-                var path = Path.Combine(Directory.GetCurrentDirectory(), FILES_DIRECTORY);
-                if (!Directory.Exists(path))
+                var file = _context.UserFiles.FirstOrDefault(f => f.Id == id);
+                if (file != null)
                 {
-                    return BadRequest("File not found");
+                    var systemName = file.SystemName;
+                    _context.UserFiles.Remove(file);
+                    _context.SaveChanges();
+
+                    var path = Path.Combine(Directory.GetCurrentDirectory(), FILES_DIRECTORY);
+                    if (!Directory.Exists(path))
+                    {
+                        return BadRequest("File not found in file system");
+                    }
+                    System.IO.File.Delete(Path.Combine(path, systemName));
+
+                    return Ok();
                 }
-                System.IO.File.Delete(Path.Combine(path, systemName));
-                return Ok();
+                else
+                {
+                    return BadRequest("File not found in data base");
+                }
             }
             catch (Exception ex)
             {
